+++ packages/react-zeugma/src/entities/pane/ui/Tab.tsx
@@ -6,6 +6,7 @@ import {
   useZeugmaDrag,
   PortalRegistryContext,
 } from '../../../shared'
+import { useTabMetadata } from '../../zeugma/model'
 import { TabsContext } from './Tabs'
 
 export interface TabContextValue {
@@ -35,6 +36,7 @@ export const useTabContext = () => {
 export interface TabRenderProps {
   isDragging: boolean
   isOver: boolean
+  metadata?: Record<string, unknown>
 }
 
 export interface TabProps {
@@ -60,6 +62,7 @@ export const Tab: React.FC<TabProps> = ({ id, locked = false, children, classNam
   const actions = useContext(ZeugmaActionsContext)
   const { overTabId } = useZeugmaDrag()
   const portalRegistry = useContext(PortalRegistryContext)
+  const metadata = useTabMetadata(id)
 
   useEffect(() => {
     if (portalRegistry?.registerTabHeader) {
@@ -111,7 +114,6 @@ export const Tab: React.FC<TabProps> = ({ id, locked = false, children, classNam
   const separator = showSeparator ? <div className={classNames.tabSeparator} /> : null
 
   const isActive = tabsContext ? tabsContext.activeTabId === id : false
-  const metadata = tabsContext?.tabsMetadata?.[id]
 
   const selectThisTab = useCallback(() => {
     tabsContext?.selectTab(id)
@@ -168,6 +170,7 @@ export const Tab: React.FC<TabProps> = ({ id, locked = false, children, classNam
         {children({
           isDragging,
           isOver: isTargetOver,
+          metadata,
         })}
       </div>
     </TabContext.Provider>